backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
throttled-py==2.0.2
openai==1.88.0
requests==2.31.0
python-json-logger==2.0.7
boto3==1.40.69
python-dotenv==1.0.0